Update google test

+++ core/unit_test/TestReduce.hpp
@@ -585,9 +585,9 @@ TEST(TEST_CATEGORY, int_combined_reduce) {
                           Kokkos::RangePolicy<TEST_EXECSPACE>(0, nw),
                           functor_type(nw), result1, result2, result3);
 
-  ASSERT_EQ(nw, result1);
-  ASSERT_EQ(nsum, result2);
-  ASSERT_EQ(nsum, result3);
+  ASSERT_EQ(nw, uint64_t(result1));
+  ASSERT_EQ(nsum, uint64_t(result2));
+  ASSERT_EQ(nsum, uint64_t(result3));
 }
 
 TEST(TEST_CATEGORY, mdrange_combined_reduce) {
@@ -606,9 +606,9 @@ TEST(TEST_CATEGORY, mdrange_combined_reduce) {
                                                              {{nw, 1, 1}}),
       functor_type(nw), result1, result2, result3);
 
-  ASSERT_EQ(nw, result1);
-  ASSERT_EQ(nsum, result2);
-  ASSERT_EQ(nsum, result3);
+  ASSERT_EQ(nw, uint64_t(result1));
+  ASSERT_EQ(nsum, uint64_t(result2));
+  ASSERT_EQ(nsum, uint64_t(result3));
 }
 
 TEST(TEST_CATEGORY, int_combined_reduce_mixed) {
@@ -629,9 +629,9 @@ TEST(TEST_CATEGORY, int_combined_reduce_mixed) {
                           functor_type(nw), result1_v, result2,
                           Kokkos::Sum<int64_t, Kokkos::HostSpace>{result3_v});
 
-  ASSERT_EQ(nw, result1_v());
-  ASSERT_EQ(nsum, result2);
-  ASSERT_EQ(nsum, result3_v());
+  ASSERT_EQ(int64_t(nw), result1_v());
+  ASSERT_EQ(int64_t(nsum), result2);
+  ASSERT_EQ(int64_t(nsum), result3_v());
 }
 #endif
 }  // namespace Test
